pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "faultray"
version = "1.2.0"
description = "FaultRay — Zero-risk infrastructure chaos engineering. Prove your availability ceiling mathematically."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Yutaro Maeda", email = "mattyopon@gmail.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Networking",
    "Typing :: Typed",
]
keywords = ["chaos-engineering", "sre", "devops", "infrastructure", "simulation", "reliability", "slo"]
dependencies = [
    "typer>=0.15",
    "rich>=13.0",
    "pydantic>=2.0",
    "networkx>=3.0",
    "psutil>=6.0",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "jinja2>=3.1",
    "httpx>=0.28",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
]

[project.optional-dependencies]
aws = [
    "boto3>=1.28",
]
gcp = [
    "google-cloud-compute>=1.0",
    "google-cloud-sql-admin>=1.0",
    "google-cloud-redis>=2.0",
    "google-cloud-storage>=2.0",
]
azure = [
    "azure-identity>=1.15",
    "azure-mgmt-compute>=30.0",
    "azure-mgmt-sql>=3.0",
    "azure-mgmt-redis>=14.0",
]
k8s = [
    "kubernetes>=28.0",
]
all-clouds = [
    "boto3>=1.28",
    "google-cloud-compute>=1.0",
    "azure-identity>=1.15",
    "kubernetes>=28.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.urls]
Homepage = "https://github.com/mattyopon/infrasim"
Documentation = "https://github.com/mattyopon/infrasim#readme"
Repository = "https://github.com/mattyopon/infrasim"
Issues = "https://github.com/mattyopon/infrasim/issues"
Changelog = "https://github.com/mattyopon/infrasim#changelog"

[project.scripts]
faultray = "infrasim.cli:app"
chaosproof = "infrasim.cli:app"
infrasim = "infrasim.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/infrasim"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
concurrency = ["thread", "greenlet"]
source = ["src/infrasim"]
